Some cleanup to have one function per file
