+++ packages/core/services/fetchComponents.test.js
@@ -1,92 +1,53 @@
 import fetchMock from 'fetch-mock';
-import {
-  createQueryString,
-  fetchComponents,
-} from './fetchComponents';
+import fetchComponents from './fetchComponents';
 
 beforeEach(() => {
   process.env.API_ROOT_URL = 'https://foo.com/api';
   fetchMock.restore();
 });
 
-it('should build a correct query string', () => {
-  const basicQueryString = createQueryString(
-    '/basic-path'
-  );
-  const pathWithSearch = createQueryString(
-    '/some-path',
-    's=searchstring'
-  );
-  const pathWithCookie = createQueryString(
-    '/some-path',
-    '',
-    {
-      someCookieKey: 'someCookieValue',
-    }
-  );
-  const pathWithContext = createQueryString(
-    '/some-path',
-    '',
-    {},
-    'site'
-  );
-  const pathWithEveryOption = createQueryString(
-    '/some-path',
-    's=searchstring',
-    {
-      aCookie: 99,
-    },
-    'site'
-  );
+it(
+  'should append extra query params that are defined',
+  async (done) => {
+    process.env.API_QUERY_PARAM_BAR = 'baz';
 
-  expect(basicQueryString).toBe(
-    'path=/basic-path&context=page'
-  );
-  expect(pathWithSearch).toBe(
-    'path=/some-path&context=page&s=searchstring'
-  );
-  expect(pathWithCookie).toBe(
-    'path=/some-path&context=page&someCookieKey=someCookieValue'
-  );
-  expect(pathWithContext).toBe(
-    'path=/some-path&context=site'
-  );
-  expect(pathWithEveryOption).toBe(
-    'path=/some-path&context=site&s=searchstring&aCookie=99'
-  );
-});
+    // Throws an error if the request doesn't match.
+    fetchMock.get(
+      'https://foo.com/api/components',
+      {},
+      {
+        query: {
+          path: '/foo',
+          context: 'page',
+          bar: 'baz',
+        },
+      }
+    );
 
-it('should append extra query params that are defined', async (done) => {
-  process.env.API_QUERY_PARAM_BAR = 'baz';
+    const result = await fetchComponents('/foo');
 
-  // Throws an error if the request doesn't match.
-  fetchMock.get(
-    'https://foo.com/api/components',
-    {},
-    {
+    expect(result).toBeDefined();
+
+    done();
+  }
+);
+
+it(
+  'should pass query params from the request to the api',
+  async (done) => {
+    // Throws an error if the request doesn't match.
+    fetchMock.get('https://foo.com/api/components', {}, {
       query: {
         path: '/foo',
         context: 'page',
         bar: 'baz',
       },
-    }
-  );
+    });
 
-  await fetchComponents('/foo');
+    const result = await fetchComponents('/foo', '?bar=baz');
 
-  done();
-});
+    expect(result).toBeDefined();
 
-it('should pass query params from the request to the api', async (done) => {
-  // Throws an error if the request doesn't match.
-  fetchMock.get('https://foo.com/api/components', {}, {
-    query: {
-      path: '/foo',
-      context: 'page',
-      bar: 'baz',
-    },
-  });
-
-  await fetchComponents('/foo', '?bar=baz');
-  done();
-});
+    done();
+  }
+);
